+++ lib/mix/tasks/crucible.shutdown.ex
@@ -1,8 +1,0 @@
-defmodule Mix.Tasks.Crucible.Shutdown do
-  @shortdoc "Destroy a machine"
-  @moduledoc "mix crucible.shutdown --driver NAME --id ID"
-  use Mix.Task
-
-  @impl Mix.Task
-  def run(args), do: Mix.Tasks.Crucible.run(["shutdown" | args])
-end